pallet-vesting: Move `tests` module to seperate file
